Refactor mute rule functionality into seperate file
